# Refactor `window`